test: add tests for Hooks API ACL
